Resolves conflicts introduced by main's ParticipantsController/WebHosting
refactors (thin controller with ParticipantService, IWebHosting interface).
Kept main's structure and reapplied this branch's fixes on top: replaced
StopWebHosting's .Wait() with .GetAwaiter().GetResult(), and moved the
null-to-empty-list fix for GetParticipantsOfSession into ParticipantService,
since the query logic now lives there. Updated the matching test and
verified the full suite passes.

+++ F1Server.Core/Packets/Data/F1-2020/CarStatusData2020.cs
@@ -87,7 +87,7 @@ public class CarStatusData2020 : ICarStatusData2020
     /// <summary>
     /// Pit limiter status - 0 - off - 1 - on
     /// </summary>
-    public uint PitLimiterStatus { get; set; }
+    public ushort PitLimiterStatus { get; set; }
 
     /// <summary>
     /// Current fuel mass
@@ -145,7 +145,7 @@ public class CarStatusData2020 : ICarStatusData2020
     public float ERSStoreEnergy { get; set; }
 
     /// <summary>
-    /// ERS deployment mode - 0 - none - 1 - medium - 2 - hotlap - 3 - overtake
+    /// ERS deployment mode - 0 - none - 1 - medium - 2 - overtake - 3 - hotlap
     /// </summary>
     public ushort ERSDeployMode { get; set; }
