+++ F1Server.Core/Packets/Data/F1-2024/CarStatusData2024.cs
@@ -33,7 +33,7 @@ public class CarStatusData2024 : ICarStatusData2024
     /// <summary>
     /// Pit limiter status - 0 - off - 1 - on
     /// </summary>
-    public uint PitLimiterStatus { get; set; }
+    public ushort PitLimiterStatus { get; set; }
 
     /// <summary>
     /// Current fuel mass
